Refactor generate factory components and update example prompts
- Removed unused Flutter material import in provider.dart.
- Updated example prompts in GenerateFactoryState to reflect new labels and descriptions.
- Adjusted layout in generate_factory_modal_step1.dart to ensure error messages are displayed correctly.
- Modified generate_factory_modal_step3.dart to show a close button when the factory is created.
- Cleaned up imports in generate_factory_modal.dart.
- Removed redundant loading state handling in training_session provider.
- Enhanced record panel functionality to navigate to a recording overlay view upon starting a recording.
- Updated DemonstrationReward model to change max_reward type from i64 to f64 for better precision.

+++ src-tauri/src/core/record.rs
@@ -61,7 +61,7 @@ pub struct Demonstration {
 #[derive(Serialize, Deserialize, Clone)]
 pub struct DemonstrationReward {
     time: i64,
-    max_reward: i64,
+    max_reward: f64,
 }
 
 /// Information about the primary monitor used for recording.
